extracted logic of punishment removal into punishment service
